Added IO metrics, tied analysers with container IDs

+++ analysers/trials/processDuration.py
@@ -93,11 +93,11 @@ q3 = np.percentile(data, 75).item()
 p95 = np.percentile(data, 95).item()
 
 #mean = reduce(lambda x, y: x + y, data) / float(dataLength)
-mean = np.mean(data).item()
+mean = np.mean(data, dtype=np.float64).item()
 #variance = map(lambda x: (x - mean)**2, data)
-variance = np.var(data).item()
+variance = np.var(data, dtype=np.float64).item()
 #stdD = math.sqrt(sum(variance) * 1.0 / dataLength)
-stdD = np.std(data).item()
+stdD = np.std(data, dtype=np.float64).item()
 
 stdE = stdD/float(math.sqrt(dataLength))
 marginError = stdE * 2
